requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pinecone>=2.2.1
sentence-transformers>=2.2.0
python-dotenv>=0.21.0
python-multipart>=0.0.5
numpy>=1.22.0
pydantic>=2.0.0
setuptools>=65.0.0
